# backend/openapi.yaml
openapi: 3.0.3
info:
  title: УК-инциденты API
  version: 0.3.0
  description: |
    Бэкенд мини-аппа MAX для жильцов и УК. Источник правды по контракту —
    этот файл: из него генерируется Go-сервер (`make gen`) и TS-клиент для фронта
    (`npx openapi-typescript backend/openapi.yaml -o src/shared/api/schema.d.ts`).

    Контракт зафиксирован фронтом (shared/api/client.ts + zod-схемы): camelCase,
    nullable-поля всегда присутствуют в ответе (ключ есть, значение null),
    ошибки — единая форма {code, message, details?}.

    Авторизация: заголовок `Authorization: tma <initData>`. `auth_date` не старше 24ч.
    Ручки `/api/uk/*` и `PATCH /api/incidents/{id}/status` — только для роли
    `uk_dispatcher` (иначе 403 forbidden).
servers:
  - url: http://localhost:8080
security:
  - TmaInitData: []
paths:
...
  /api/photos/{id}:
    get:
      operationId: getPhoto
      security: []
      tags: [incidents]
      summary: Содержимое фото (публично — вебвью грузит <img> без заголовка Authorization)
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200":
          description: Файл
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/Error" }
  /api/incidents/{id}/status:
    patch:
      operationId: setIncidentStatus
      tags: [uk]
      summary: "Смена статуса диспетчером: accepted → in_progress → verifying → done"
      description: |
        Переход в `verifying` рассылает подписчикам «починили — подтвердите»;
        переход в `done` проставляет resolvedAt. Любой другой переход — 422.
      parameters: [{ $ref: "#/components/parameters/ID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetIncidentStatusRequest" }
      responses:
        "200":
          description: Авария после перехода
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/uk/queue:
    get:
      operationId: ukQueue
      tags: [uk]
      summary: Очередь незакрытых заявок по домам УК диспетчера (critical сначала, затем createdAt desc)
      parameters:
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Страница очереди
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedUkQueue" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/uk/events:
    post:
      operationId: ukCreateEvent
      tags: [uk]
      summary: Создать плановое событие (дом должен принадлежать УК диспетчера, иначе 404)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventRequest" }
      responses:
        "201":
          description: Событие создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    TmaInitData:
      type: apiKey
      in: header
      name: Authorization
      description: "`tma <initData>` из MAX WebApp Bridge"
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorResponse" }
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: validation_failed }
        message: { type: string }
        details: {}
...
    Severity:
      type: string
      enum: [critical, warning]
    IncidentStatus:
      type: string
      enum: [accepted, in_progress, verifying, done]
    Incident:
      type: object
      required: [id, houseId, title, description, severity, status, affectedCount,
                 createdAt, dueAt, joinedByMe, confirmedByMe, photos]
      properties:
        id: { type: string, example: inc_01H }
        houseId: { type: string, example: h_123 }
        title: { type: string }
        description: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        status: { $ref: "#/components/schemas/IncidentStatus" }
        affectedCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time, nullable: true }
        joinedByMe: { type: boolean }
        confirmedByMe: { type: boolean }
        photos:
          type: array
          items: { $ref: "#/components/schemas/Photo" }
    Photo:
      type: object
      required: [id, url]
      properties:
        id: { type: string, example: ph_1 }
        url: { type: string, example: /api/photos/ph_1 }
    SetIncidentStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/IncidentStatus" }
...
    EventStatus:
      type: string
      enum: [planned, in_progress, awaiting_confirmation, disputed, closed]
    Event:
      type: object
      required: [id, houseId, entrance, riser, reason, responsible, scheduledFrom, scheduledTo,
                 status, createdAt, resolvedAt]
      properties:
        id: { type: string, example: evt_1 }
        houseId: { type: string, example: h_1 }
        entrance: { type: string, nullable: true }
        riser: { type: string, nullable: true }
        reason: { type: string, example: Плановое отключение горячей воды }
        responsible: { type: string, example: УК-1 }
        scheduledFrom: { type: string, format: date-time }
        scheduledTo: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/EventStatus" }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
...
    CreateEventRequest:
      type: object
      required: [houseId, reason, responsible, scheduledFrom, scheduledTo]
      properties:
        houseId: { type: string, example: h_1 }
        entrance: { type: string }
        riser: { type: string }
        reason: { type: string, minLength: 1 }
        responsible: { type: string, minLength: 1 }
        scheduledFrom: { type: string, format: date-time }
        scheduledTo: { type: string, format: date-time }
    UkQueueItem:
      type: object
      required: [id, houseId, houseAddress, title, description, severity, status, createdAt,
                 dueAt, affectedCount, confirmedCount, reporterName]
      properties:
        id: { type: string, example: inc_42 }
        houseId: { type: string, example: h_1 }
        houseAddress: { type: string, example: "Казань, ул. Баумана, д. 10" }
        title: { type: string }
        description: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        status: { $ref: "#/components/schemas/IncidentStatus" }
        createdAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time, nullable: true }
        affectedCount: { type: integer, minimum: 0 }
        confirmedCount: { type: integer, minimum: 0 }
        reporterName: { type: string, example: Иван И. }
    PaginatedUkQueue:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/UkQueueItem" }
        total: { type: integer, minimum: 0 }
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 0 }
...